[Shaman] Implement Flowing Spirits random distribution
It looks like in-game, there's a diminishing proc chance the more procs
(either 1 or 2 wolves) you have up at a time. Experimentally verified to
correspond to roughly, 6.65%, 4.61%, 2.36%, 1.25% proc chance per active
proc, starting from 0, for a maximum of 4/8 Spirit Wolves up
simulatenously, depending on the TWW set bonus.

Note that while the experimentation did not witness successful procs
past 4 active procs, this does not confirm that the maximum active
number of procs is 4; however, the effect of 5 simultaneous procs to the
simulation is negligible.

This commit also removes the old Flowing Spirits-related options, and
introduces a new one. The `shaman.flowing_spirit_proc_chances` option
accepts a ':'-delimited set of proc chances for the spirit wolves. The
values are either from the range [0..1], or >1 in which case the value
will be divided by 100. This option allows the user to bypass the 4
active proc limit if they so choose, and also allows the user to reduce
the limit down to a maximum of 1.