[Core] Fix precombat order of operations (#9781)
* Fix precombat order of operations
Fixes the order of event issues present due to the Singing Citrine effects. In order to accomplish this, i've removed the APL action accessibility for Flask, Food and Augmentation actions and instead trigger them early in `player_t::arise()`. This is a requirement due to the APL actions being called after precombat_callbacks that are often utilized to apply effects that will utilize the player's current state on combat entry, which previously did not include flask, food, and augment rune effects. Due to these effects lasting through death, they should already be present on the player whenever buffs are applied which is accomplished with this change.  

May also change the results with effects like Ovinax's Mercurial Egg, if the stat provided by flask or food changes the highest secondary stat, as before this wouldn't be updated until the first tick.